update: change name & abstract function
apiOuter.js -> dappsApi.js
abstract dapps api
